put deviceId back to being a long to make the protocol validator happy

+++ src/main/java/io/finn/signald/clientprotocol/v1/RemoveLinkedDeviceRequest.java
@@ -26,14 +26,14 @@ import org.whispersystems.signalservice.api.push.exceptions.AuthorizationFailedE
 public class RemoveLinkedDeviceRequest implements RequestType<Empty> {
   @ExampleValue(ExampleValue.LOCAL_PHONE_NUMBER) @Doc("The account to interact with") @Required public String account;
 
-  @ExampleValue("3") @Doc("the ID of the device to unlink") @Required public int deviceId;
+  @ExampleValue("3") @Doc("the ID of the device to unlink") @Required public long deviceId;
 
   @Override
   public Empty run(Request request) throws InternalError, InvalidProxyError, ServerNotFoundError, NoSuchAccountError, InvalidRequestError, AuthorizationFailedError {
     Account a = Common.getAccount(account);
     SignalServiceAccountManager accountManager = Common.getDependencies(a.getUUID()).getAccountManager();
     try {
-      accountManager.removeDevice(deviceId);
+      accountManager.removeDevice((int)deviceId);
     } catch (AuthorizationFailedException e) {
       throw new AuthorizationFailedError(e);
     } catch (IOException e) {
